Se cambia el atributodel coeficiente de int a float para la operacion de division. Se implementaron las operaciones de multiplicar y dividir. Se mejoro la visualizacion de los polinomios insertados y el resultado de las operaciones de estos. Se agrego el boton vaciar polinomios para reiniciar los polinomios. Al parecer proyecto terminado.

+++ src/polinomio/Termino.java
@@ -2,23 +2,23 @@ package polinomio;
 
 public class Termino {
     
-    private int coeficiente;
+    private float coeficiente;
     private int exponente;
     
     public Termino(){
         
     }
 
-    public Termino(int coeficiente, int exponente) {
+    public Termino(float coeficiente, int exponente) {
         this.coeficiente = coeficiente;
         this.exponente = exponente;
     }
 
-    public int getCoeficiente() {
+    public float getCoeficiente() {
         return coeficiente;
     }
 
-    public void setCoeficiente(int coeficiente) {
+    public void setCoeficiente(float coeficiente) {
         this.coeficiente = coeficiente;
     }
